--- pkg/cmd/server/apis/config/validation/build_overrides.go
@@ -1,18 +1,0 @@
-package validation
-
-import (
-	"k8s.io/apimachinery/pkg/util/validation/field"
-	"k8s.io/kubernetes/pkg/apis/core/validation"
-
-	buildvalidation "github.com/openshift/origin/pkg/build/apis/build/validation"
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
-)
-
-func ValidateBuildOverridesConfig(config *configapi.BuildOverridesConfig) field.ErrorList {
-	allErrs := field.ErrorList{}
-	allErrs = append(allErrs, buildvalidation.ValidateImageLabels(config.ImageLabels, field.NewPath("imageLabels"))...)
-	allErrs = append(allErrs, buildvalidation.ValidateNodeSelector(config.NodeSelector, field.NewPath("nodeSelector"))...)
-	allErrs = append(allErrs, validation.ValidateAnnotations(config.Annotations, field.NewPath("annotations"))...)
-
-	return allErrs
-}
