--- pkg/cmd/server/apis/config/validation/validation_test.go
@@ -1,1 +1,0 @@
-package validation
