+++ pkg/oc/cli/admin/groups/sync/sync.go
@@ -24,9 +24,9 @@ import (
 	"github.com/openshift/library-go/pkg/security/ldapclient"
 	"github.com/openshift/origin/pkg/cmd/server/apis/config"
 	configapilatest "github.com/openshift/origin/pkg/cmd/server/apis/config/latest"
-	"github.com/openshift/origin/pkg/cmd/server/apis/config/validation/ldap"
 	"github.com/openshift/origin/pkg/oc/lib/groupsync"
 	"github.com/openshift/origin/pkg/oc/lib/groupsync/interfaces"
+	"github.com/openshift/origin/pkg/oc/lib/groupsync/ldap"
 	"github.com/openshift/origin/pkg/oc/lib/groupsync/syncerror"
 )
